move logic to service and update on meter changes
