+++ src/app/upload-data/data-setup/file-setup/confirm-and-submit/confirm-and-submit.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-confirm-and-submit',
+  templateUrl: './confirm-and-submit.component.html',
+  styleUrls: ['./confirm-and-submit.component.css']
+})
+export class ConfirmAndSubmitComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
