+++ src/app/upload-data/upload-data.module.ts
@@ -17,6 +17,7 @@ import { TemplateFacilitiesComponent } from './data-setup/file-setup/template-fa
 import { EditMeterFormModule } from '../facility/utility-data/energy-consumption/energy-source/edit-meter-form/edit-meter-form.module';
 import { ConfirmReadingsComponent } from './data-setup/file-setup/confirm-readings/confirm-readings.component';
 import { ConfirmPredictorsComponent } from './data-setup/file-setup/confirm-predictors/confirm-predictors.component';
+import { ConfirmAndSubmitComponent } from './data-setup/file-setup/confirm-and-submit/confirm-and-submit.component';
 
 
 
@@ -34,7 +35,8 @@ import { ConfirmPredictorsComponent } from './data-setup/file-setup/confirm-pred
     SetFacilityMetersComponent,
     TemplateFacilitiesComponent,
     ConfirmReadingsComponent,
-    ConfirmPredictorsComponent
+    ConfirmPredictorsComponent,
+    ConfirmAndSubmitComponent
   ],
   imports: [
     CommonModule,
